requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
selenium
fuzzywuzzy[speedup]
